+++ cmd/samples/pageflow/src/server/helpers/handleError.js
@@ -1,0 +1,5 @@
+const handleError = ({ error, response }) => {
+  return response.code(error.code || '500').send(JSON.stringify(error));
+};
+
+export default handleError;
