+++ cmd/samples/pageflow/src/server/helpers/momentToLong.js
@@ -1,0 +1,5 @@
+import Long from 'long';
+
+const momentToLong = m => Long.fromValue(m.unix()).mul(1000000000);
+
+export default momentToLong;
