+++ cmd/samples/pageflow/src/server/helpers/parseContent.js
@@ -1,0 +1,8 @@
+const parseContent = (str) => {
+  console.log('str:', str);
+  const parsedString = str.replace(/^"/, '').replace(/"$/, '').replace(/\\/g, '');
+  console.log('parsing:', parsedString);
+  return JSON.parse(parsedString);
+};
+
+export default parseContent;
