+++ cmd/samples/pageflow/src/server/helpers/updateProductDescription.js
@@ -1,0 +1,9 @@
+const updateProductDescription = ({ description, id }) => {
+  // TODO - communicate with cadence-server to fetch product information from workflow.
+
+  const product = getProduct(id);
+  product.description = description;
+  return product;
+};
+
+export default updateProductDescription;
