+++ cmd/samples/pageflow/src/server/helpers/waitTime.js
@@ -1,0 +1,2 @@
+const waitTime = (time) => new Promise((resolve) => setTimeout(resolve, time));
+export default waitTime;
